backend/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
pydantic
python-dotenv
